requirements.txt
aiosqlite==0.19.0
alembic==1.11.1
anyio==3.6.2
appnope==0.1.3
apprise==1.4.0
asgi-lifespan==2.1.0
asttokens==2.2.1
asyncpg==0.27.0
attrs==23.1.0
azure-core==1.26.4
azure-storage-blob==12.16.0
backcall==0.2.0
botocore==1.29.135
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
coolname==2.2.0
croniter==1.3.14
cryptography==40.0.2
dateparser==1.1.8
debugpy==1.6.7
decorator==5.1.1
delta-spark==2.3.0
docker==6.1.2
executing==1.2.0
fastapi==0.95.2
fsspec==2023.5.0
google-auth==2.18.1
greenlet==2.0.2
griffe==0.28.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.1
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.1
ipython==8.13.2
isodate==0.6.1
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kubernetes==26.1.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.24.3
oauthlib==3.2.2
orjson==3.8.12
packaging==23.1
pandas==2.0.1
parso==0.8.3
pathspec==0.11.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prefect==2.10.10
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
pyrsistent==0.19.3
pyspark==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==25.0.2
readchar==4.0.5
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
stack-data==0.6.2
starlette==0.27.0
text-unidecode==1.3
toml==0.10.2
tornado==6.3.2
traitlets==5.9.0
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.15
uvicorn==0.22.0
wcwidth==0.2.6
websocket-client==1.5.1
websockets==11.0.3
zipp==3.15.0
